feat(web): enable Cache Components / PPR (closes #2835, #2826)
Flips `cacheComponents: true` in next.config.ts and migrates the routes
that were incompatible with the new model. Public marketing pages and
the 4 ISR pages now serve as `◐ Partial Prerender` — static shell + cached
body baked into the initial HTML, with dynamic subtrees streamed via
Suspense. Bottom line: AI fetchers and JS-disabled crawlers now see the
real page body in the initial response instead of just the metadata head
and a thin React shell with hydration markers.

Highlights of the migration:

- **Approach A for `<html lang>` per locale (#2826).** Deletes
  `app/layout.tsx` and `app/loading.tsx` entirely. `app/[lang]/layout.tsx`
  becomes the de-facto root layout — owns `<html lang={locale}>`, `<body>`,
  ThemeProvider, Vercel Analytics + Speed Insights, globals.css import,
  and the `metadata` export. Locale comes from `params` (deterministic
  under `generateStaticParams`), so the static shell stays static — zero
  `headers()` reads, zero per-request function invocations for the
  language attribute. Drops the inline `langScript` client-side rewriter.

- **Auth layout Suspense pattern.** Splits the session check off the
  layout's render path. Static `AuthShell` + form prerender; a separate
  `RedirectIfSignedIn` async child runs inside `<Suspense fallback={null}>`,
  reads `headers()` + Better Auth, and `redirect()`s if signed in. Auth
  pages are inherently dynamic anyway, so the dynamic hole is acceptable.

- **Page-level `revalidate = N` migrated to `'use cache'` + `cacheLife()`.**
  The 4 ISR pages (explore, company/[slug], [user]/[watchlist], blog)
  plus their `generateMetadata` functions now declare `'use cache'` and
  `cacheLife({ revalidate: N })` instead of the route-segment export
  (which the build now rejects). Same revalidate semantics, but the
  cached body is baked into the prerender HTML.

- **OG image route configs removed.** `next/og` `ImageResponse` isn't
  serializable for `'use cache'`, but Next.js caches OG images via HTTP
  Cache-Control headers automatically. Removing the (now-rejected)
  `revalidate = 2592000` exports is sufficient.

- **Other route-segment configs removed:** `runtime = "nodejs"` from 6
  API routes (Fluid Compute Node.js is the default anyway),
  `dynamic = "force-dynamic" + revalidate = 0` from `indexnow-key.txt`,
  `dynamicParams = false` from blog post page (notFound() handles missing
  slugs).

- **Footer copyright year** moved to module scope. `new Date()` in a
  server-render path is rejected; module-scope evaluation refreshes on
  every deploy, which is enough for a copyright year.

- **Custom Redis `cached()` helper untouched.** Per the issue,
  cross-instance shared state and the `skipIf` empty-result predicate
  aren't replicated by `'use cache'` semantics. Strategic per-call-site
  migration is tracked as a follow-up.

- **Retired `app/__tests__/isr-routes.test.ts`.** Under cacheComponents
  the build itself enforces what the line-by-line scan was checking; a
  build-output classifier as a richer replacement is tracked as a
  follow-up.

- **`apps/web/docs/cache-components.md`** updated: rewrites the
  per-locale `<html lang>` recipe to match Approach A, expands the
  migration cheat sheet with the segment configs the build rejects, adds
  the `new Date()` and OG image patterns I learned the hard way during
  this migration.

- `pnpm build` succeeds; route summary shows every `/[lang]/*` HTML
  route as `◐ Partial Prerender` (was previously a mix of ISR + dynamic).
- `pnpm test` — 446/446 tests pass.
- `pnpm exec tsc --noEmit` — clean (per the strict-tsc workflow rule
  Vercel enforces but PR CI doesn't).
- `pnpm dev` + `curl` smoke-test on every public page across all 4
  locales: HTTP 200, `<html lang="{locale}">` server-rendered, body
  content visible in initial HTML (e.g. `/en/` returns 5 `watchlist` and
  6 `career` matches in the body, no JS execution required).
- Playwright walkthrough (en/de/fr/it × home/about/faq/how-we-index/blog,
  plus sign-in, sign-up, explore) — all pages render, no JS errors,
  screenshots captured.
