+++ apps/web/app/[lang]/(app)/company/[slug]/opengraph-image.tsx
@@ -5,12 +5,11 @@ import { locales } from "@/lib/i18n";
 export const alt = "Company jobs";
 export const size = { width: 1200, height: 630 };
 export const contentType = "image/png";
-// 30 days. Company logo/name/description change far less often than
-// daily — a 1-day revalidate multiplied by ~16k (slug × locale) cells
-// churned OG regenerations against the crawl surge. Bump high; a
-// rename/rebrand still propagates within a month, and deploys purge
-// the cache anyway.
-export const revalidate = 2592000;
+// Cached automatically via the Next.js OG image route's built-in
+// Cache-Control headers; no `'use cache'` needed (ImageResponse is a
+// class instance and isn't serializable for the runtime cache).
+// `generateStaticParams` below covers the top-N companies at build
+// time so social-card crawl surges land on the prebake.
 
 import { readFile } from "node:fs/promises";
 import { join } from "node:path";
