+++ apps/web/app/[lang]/(public)/blog/[slug]/page.tsx
@@ -3,6 +3,7 @@ import { notFound } from "next/navigation";
 import Link from "next/link";
 import { compileMDX } from "next-mdx-remote/rsc";
 import remarkGfm from "remark-gfm";
+import { cacheLife } from "next/cache";
 import { getI18n } from "@lingui/react/server";
 import { initI18nForPage, isLocale, defaultLocale } from "@/lib/i18n";
 import { siteConfig } from "@/content/config";
@@ -17,16 +18,11 @@ import {
 import { buildMdxComponents } from "@/components/blog/MdxMentions";
 import { RelatedPosts } from "@/components/blog/RelatedPosts";
 
-// Posts are static content authored at PR-merge cadence; ISR window is
-// cosmetic. Build-time prerender via `generateStaticParams` covers
-// every published post.
-export const revalidate = 86400;
-// English-only initially; once a post has translated MDX siblings the
-// per-post hreflang map can be widened in `generateMetadata`. For now
-// only `/en/blog/{slug}` is the canonical surface — the localized
-// paths exist (`generateStaticParams` emits all 4 locales for routing
-// completeness) but redirect intent flows through the canonical EN URL.
-export const dynamicParams = false;
+// Posts are static content authored at PR-merge cadence — build-time
+// prerender via `generateStaticParams` covers every published post.
+// Non-existent slugs fall through to `getBlogPost` returning null and
+// `notFound()` firing in the page body. (`dynamicParams = false` would
+// be the more direct route, but it's incompatible with cacheComponents.)
 
 type Props = {
   params: Promise<{ lang: string; slug: string }>;
@@ -42,6 +38,8 @@ export async function generateStaticParams(): Promise<{ lang: string; slug: stri
 }
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
+  "use cache";
+  cacheLife({ revalidate: 86400 });
   const { lang, slug } = await params;
   const locale = isLocale(lang) ? lang : defaultLocale;
   const post = await getBlogPost(slug, locale);
@@ -116,6 +114,8 @@ function formatDate(iso: string, locale: string): string {
 }
 
 export default async function BlogPostPage({ params }: Props) {
+  "use cache";
+  cacheLife({ revalidate: 86400 });
   const locale = await initI18nForPage(params);
   const i18n = getI18n()!;
   const { slug } = await params;
